requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
emoji==1.6.1
PySocks==1.7.1
python-telegram-bot==13.5
pytz==2021.1
six==1.16.0
tornado==6.1
tzlocal==2.1
emoji==1.6.1